fix: Add suspense for Build error
